requirements.txt
bibtexparser == 1.1.0
csscompressor == 0.9.5
django-compressor == 2.4
django[argon2] == 3.0.3
gunicorn == 20.0.4
psycopg2-binary == 2.8.4
pylibmc == 1.6.1
python-dotenv == 0.11.0
requests == 2.23.0
